Fix Basic attribute return type.

+++ capability/power_supply/callbacks.go
@@ -70,7 +70,7 @@ func (i *Implementation) EnumerateDevice(ctx context.Context, d zda.Device) erro
 		}
 
 		if basicResp[basic.PowerSource].Status == 0 {
-			value := basicResp[basic.PowerSource].DataTypeValue.Value.(uint8)
+			value := basicResp[basic.PowerSource].DataTypeValue.Value.(uint64)
 
 			if (value&0x80) == 0x80 || (value&0x7f) == 0x03 {
 				battery.Available = true
